дсд

+++ GamingStore2/Gaming_Store_Data/Data/OrderDto.cs
@@ -10,6 +10,6 @@ namespace Gaming_Store_Data.Data
     {
         public int Id { get; set; }
         public int GameId { get; set; }
-        public decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
     }
 }
